CM21 Integration; UI adjustment

+++ Pkg_Checker/Integration/CM21.cs
@@ -1,6 +1,7 @@
 ﻿using Pkg_Checker.Entities;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
@@ -16,29 +17,39 @@ namespace Pkg_Checker.Integration
     /// </summary>
     public class CM21
     {
+        [DefaultValue(@"C:\Program Files\honeywell_eng\cm21_v2_2b2\bin\cm21.exe")]
         public String CM21ExePath { get; set; }
         public Process CM21Process { get; set; }
+        public int ExitCode { get; set; }
 
         private const int maxWaitingTime = 3 * 60 * 1000;  // 3 min
 
-        public CM21(String exePath, String EID, String PWD, String proj, String subProj, String cwd)
+#warning Search registry for installed apps.
+        public CM21(String EID, String PWD, String proj, String subProj, String outputPath)
         {
-            if (String.IsNullOrWhiteSpace(exePath) || !File.Exists(exePath))
-            {
-                String defaultPath = @"C:\Program Files\honeywell_eng\cm21_v2_2b2\bin\cm21.exe";
-                if (File.Exists(defaultPath))
-                    CM21ExePath = defaultPath;
-                else
-                    throw new FileNotFoundException("CM21 executable is not found.");
-            }
-            else
-                CM21ExePath = exePath;
+            CM21ExePath = @"C:\Program Files\honeywell_eng\cm21_v2_2b2\bin\cm21.exe";
+            if (!File.Exists(CM21ExePath))               
+               // query registry
+               throw new FileNotFoundException("CM21 executable is not found.");           
 
             CM21Process = new Process();
             CM21Process.StartInfo.FileName = CM21ExePath;
-            CM21Process.StartInfo.Arguments = String.Format(@"-user={0} -pass={1} -proj={2} -sub={3} -dir={4}", EID, PWD, proj, subProj, cwd);
+            CM21Process.StartInfo.Arguments = String.Format(@"-user={0} -pass={1} -proj={2} -sub={3} -def={4}",
+                EID, PWD, proj, subProj, outputPath);
             CM21Process.StartInfo.WindowStyle = ProcessWindowStyle.Hidden;
             CM21Process.Start();
+            if (CM21Process.WaitForExit(1))
+            {
+                // ExitCode == 0 means logged in successfully
+                if (CM21Process.HasExited && 0 != CM21Process.ExitCode)
+                    throw new ApplicationException(String.Format("CM21 login failed with exit code {0}.", CM21Process.ExitCode));
+            }
+            else
+            {
+                if (!CM21Process.HasExited)
+                    CM21Process.Kill();  // HasExited property will become true after calling Kill()
+                throw new ApplicationException("CM21 process is stuck or exited unexpectly.");
+            }
         }
 
          ~CM21()
@@ -47,26 +58,35 @@ namespace Pkg_Checker.Integration
         }
 
         // report scr/p1=SCR_NUMBER
-        public void FetchSCRReport(List<SCRReport> reports)
-        {
-            foreach (SCRReport report in reports)
-            {
-                CM21Process.StartInfo.Arguments = String.Format("REPORT SCR/P1=\"{0}\"/OUTPUT={1}", report.SCRNumber, report.SCRNumber);
-                CM21Process.Start();
-                if (!CM21Process.WaitForExit(maxWaitingTime))
-                    CM21Process.Kill();
-            }
+        public void FetchSCRReport(IEnumerable<SCRReport> reports)
+        {            
+           foreach (SCRReport report in reports)
+           {
+               // CM21Process.HasExited: is previous command has finished?
+               if (null != CM21Process && CM21Process.HasExited)
+               {
+                   String scrNumber = report.SCRNumber.ToString("0.00");
+                   CM21Process.StartInfo.Arguments = String.Format("REPORT SCR/P1=\"{0}\"/OUTPUT={1}", scrNumber, scrNumber);
+                   CM21Process.Start();
+                   if (!CM21Process.WaitForExit(maxWaitingTime))
+                       CM21Process.Kill();
+               }
+           }            
         }
-
 
         // wait for process to exit: (Process.Exit event)
         // Process.Exit += () => {}
         public void Exit()
         {
-            CM21Process.StartInfo.Arguments = "EXIT";
-            CM21Process.Start();
-            if (!CM21Process.WaitForExit(maxWaitingTime))
-                CM21Process.Kill();  // kill the hang process
+            if (null != CM21Process)
+            {
+                CM21Process.StartInfo.Arguments = "EXIT";
+                CM21Process.Start();
+                if (!CM21Process.WaitForExit(maxWaitingTime))
+                    CM21Process.Kill();
+                else
+                    CM21Process.Close();
+            }
         }
     }
 }
